Utility scripts to create JSON files from Unicode.org data
