+++ utils/requirements.txt
@@ -1,2 +1,3 @@
 requests>=2.32.3
 beautifulsoup4>=4.12.3
+cloudscraper>=1.2.71
